Springified controller service
